scripts/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
anthropic>=0.25.0
youtube-transcript-api>=0.6.2
scholarly>=1.7.11
python-slugify>=8.0.1
